chore: Resturcture View for HomeView
